Adds Debian 13 build workflow (#402)
* Adds Debian 13 build workflow

Adds a new GitHub Actions workflow to build and package the project for Debian 13.

The workflow installs necessary dependencies, configures CMake, builds the project, packages it into a .deb file, and uploads it as a workflow artifact.

On master branch pushes, it also uploads the package to SourceForge for nightly builds.

* Updates Debian container image to Trixie

Updates the Debian container image used for testing
to the 'trixie' release. This ensures that the tests
run against the latest stable version of Debian.

* Updates nightly build target path

Updates the target path for nightly builds on Sourceforge, incorporating the job name for better organization.

Also, updates download links and badges in README to reflect current builds.